FEAT: code cleanup
